// movie-review/my-app/types/Movie.ts
export interface Movie {
  id: number;
  title: string;
  poster_path: string;
  release_date: string;
  vote_average: number;
  overview: string;
}

  // export interface Movie {
  //   id: number;
  //   title: string;
  //   content: string;
  //   createdAt: string;
  //   imgUrl: string;
  //   rating: number;
  // }